add fitness penalty for crashing into walls; reduce scale of fitness graph; make fitness be double precision everywhere

+++ src/de/hansinator/fun/jgp/life/FitnessEvaluator.java
@@ -3,7 +3,7 @@ package de.hansinator.fun.jgp.life;
 
 public interface FitnessEvaluator
 {
-	int getFitness();
+	double getFitness();
 	
 	void setFitness(int fitness);
